Phonebook step 10
Adds functionality to update existing phonebook entries when the name
provided is the same as an existing entry. User is prompted to check if
they want to update first, if not, the execution is stopped and
nothing is changed. Otherwise, db is updated.

Adds update function to personsService for submitting PUT requests.