Bloglist expansion, step 10
Testing updated to pass-through JSON web token (JWT) for authentication.

To enable testing to perform tasks restricted for registered users a
user is created before testing is performed. The user is then logged in
and a token returned. The token is stored as a global variable,
accessible from the tests.

The userId is also stored as a global var. This is then inserted when
the new blogs are created before each test. This enables the created
user to perform add, edit and delete actions for testing.

Middleware status code updated for validation error's as it was
incorrectly returning 401 (unauthorized) instead of 400 (bad request).