Bloglist frontend, step 3
Implements ability to add new blogs when the user is logged in.

This is achieved by passing the JSON Web Token to the blogService
package. This is then sent as an authorization header when making the
POST request to create a new blog.

The backend takes the new blog object and confirms the user is valid, by
confirming the JWT. The user id is extracted from the decoded JWT and
used to attach the blog to the specific user.