Blog list frontend, step 8
Implements 'likes' button for each blog to increment the number of
likes.

This works by submitting a PUT request to the backend API with an
updated object, with the only change being +1 to likes.

The backend API for PUT requests had to be updated to prevent it from
trying to modify the user and/or id fields.