Blog list tests, step 4
Adds testing for BlogForm which is used to create a new blog.

To implement testing, unique ids have been added to the form. These are
used to select the individual elements and invoke event handlers.

addBlog has been renamed to handleCreateBlog to provide better insight
to its purpose.

BlogForm testing passes through a mock function which is used to capture
 the outputs of the event handlers. Expected that the results inputted
 directly via target.value are the outputs.