Bloglist end to end testing, step 1
Implements Cypress for testing and adds first test to check that login
form appears by default.

Additional setup work performed on the backend to create an additional
API route to enable the frontend to clear the database before each test.
 This is only accessible when Node is run as 'test'.

The first test ensures that the blog and user collections in the
database are empty before each test, giving a consistent starting point.
 The test checks that the first screen a user will see is the login
screen.