Bloglist end to end testing, step 6
Implements testing to validate that the blogs are displayed by most
likes to least likes.

Login and blog creation functions have been refactored. They are now
processed by API requests, instead of screen inputs, to speed up
setup for testing.

The Blog component has been updated to add a class name which is used to
 select all blog elements for the test implemented.

 The test works by selecting the aforementioned blogs. Once the promise
 has returned the blogs are converted to an array of elements using
 .get(). The text is extracted from each element and stored for
 comparison.

The assertion loops through the text and checks to make sure the first
index contains the text from the first index of blogTitlesByMostLikes.
This validates the ordering is appropriate.