Better anecdotes, step 10
Creates and implements a real-time filter.

The Filter component renders within the App, which calls the
action creator setFilter (within the reducer). This sets the filter
value. The list of anecdotes is updated when the app is re-render, with
the JS filter function keeping only those that contain the relevant
text.

A filter reducer has been created, added to the store and embedded
within the AnecdoteList component.