Anecdotes and the backend, step 2
Updates process for adding new anecdotes to store them in the backend,
as well as the store.

This required the creation of an axios post request function, which
sends the object in the required format as a POST request.

New anecdote objects are sent to the backend prior to being added to
the redux store.